Enhance Turing machine implementation with new features and optimizations - Add NONE direction to Direction enum - Modify TuringMachine to require explicit start and accept states - Adjust step count handling to use long in TuringMachineRunner and related classes - Fix tape content retrieval in Tape class

+++ src/turingmachine/TMResult.kt
@@ -12,7 +12,7 @@ package turingmachine
 data class TMResult(
     val accepted: Boolean,
     val finalState: State,
-    val steps: Int,
+    val steps: Long,
     val result: String,
     val tapeContent: String,
     val headPosition: Int
